LSP: Fix lagging behind "documentHighlight"
We highlight the current symbol on cursor changes
(="documentHighlight"), but when the contents change the text editor
widget first informs us about the cursor change, then the content
change. This resulted in the highlighting being done on the old content.
Delay the symbol highlight by a short amount of time, so it always
operates on the new content.

Change-Id: Ic24b2a853697c588ae73b102868e8f74e4437606
Reviewed-by: David Schulz <david.schulz@qt.io>