Fix LLDB gets the wrong bitfield value.
The bitpos used by lldbbridge.py:listMembers and dumper.py:extractField are
not absulate offset. If you use a bitfield in C/C++, you may not get the
correct value (for struct or class members), I modify the bitpos in those
functions to an absolute offset, maybe it works fine.

Signed-off-by: corvofeng <corvofeng@gmail.com>