TextEditor: Cleanup text styles
So far there were different settings for different mix-ins. There was no
technical reason why we limit the user. So you can set now the color, the
relative saturation and value, the font, underline, italic and bold. The
relative saturation and value are disabled if the color is set. The
settings are reordered too so the order reflects the semantic meaning much
better.

Change-Id: I440967d6729f720816d2bc028ff9fe8e8868074e
Task-number: QTCREATORBUG-21105
Reviewed-by: Ivan Donchevskii <ivan.donchevskii@qt.io>