ProjectExplorer: Make kits available earlier during startup
delayedInitialize is run in reverse dependency order, so
ProjectExplorer's will naturally be run very late during startup.
It also depends on which other plugins do what during delayed
initialization.
Since we want the kits to be available early during startup, so the user
can e.g. click on a project in Welcome mode and have them available,
it is better to use a custom, short timer for this instead.

Task-number: QTCREATORBUG-19381
Change-Id: I7bafa9cf77b86c11d420bb684eadf51c48abd5ed
Reviewed-by: hjk <hjk@qt.io>