Respect the setting of the version code via the pro file
As a pro file can now set the version name and code via the pro file,
this enables support for the code part as it will default to a
variable (similar to the version name). Therefore we need to use
QLineEdit for this over QSpinBox so that it is kept intact.

Done-with: Markus Maier <markus.maier@rosenberger.de>
Change-Id: I0524a90dc94888c11365164c41259a934caffa1c
Reviewed-by: Markus Maier <markus.maier@rosenberger.de>
Reviewed-by: Vikas Pachdha <vikas.pachdha@qt.io>