Fix build with Qt 5.12: fully-qualify base class
Without the full name, moc will generate invalid code. This will happen
for any base class whose name exactly matches the namespace it is in,
because in

  struct A : public B::B

inside class A, B is the injected name of the parent class, not the
namespace, which in turn means B::B is the constructor.

Change-Id: If7e743cf8476463880ccfffd155f3f51aa29e8b9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>