Utils: Decrease default maximum history size
History completer has been used now in various areas
and storing 30 items for each of them inside the can
grow the settings quite a lot.
Limit the default to a sensible value - if one of the
history completer needs a higher maximum this can be
raised from the using code.

Change-Id: I5b0bdaa74c1be29f981593f840d5605b75bd17c5
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>