winrt: Remove --install from winrtrunner's default run arguments
Passing --install to winrtrunner on run will cause the application to be
reinstalled if it was installed before. This reinstallation will result
in a clean sandbox so that all previously saved data will be lost.

If the user needs described behavior he can use the existing "Uninstall
package after application stops" aspect from the run configuration.

Change-Id: I5b5975fb4646a625bfc86975770a1860effcfc06
Fixes: QTCREATORBUG-20208
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>