macOS: Fix that Preferences do not work with german translation
By default Qt does heuristics on the menu item text, to decide the menu
role of an item, including whether the item should be used for the
Preferences entry in the application menu.

Now, "QML Profiler Options" is translated to "Einstellungen des QML-
Profilers" which matches the heuristics for the "Preferences" item,
replacing the actual Preferences.

For menu items we already force the menu role to NoRole if there hasn't
been a role explicitly set. Do that for the generated menus as well.

Change-Id: I4abb3f20fd51ef033e5ec68c2283af8b7b52927d
Reviewed-by: Christian Stenger <christian.stenger@qt.io>