Debugger: Split expressions and locals and anchor Registers to Locals
The Locals view is the first tab, and the Expressions view is not visible
at all. If you grab a variable to the Locals view, it just "disappears",
and you need to know where to look in order to find it.

Return to the previous layout, when Expressions and Locals were both
visible in the same view.

Change-Id: I775fe88beb921c61c048ba7106d05ae08b8fe026
Reviewed-by: hjk <hjk@qt.io>