Debugger: Fix conflict of "Start and Break on Main" and "Step Over"
After the very first start of a debugger they were both enabled.
Make sure that the state at startup is the same as the state when
switching to "Debugger Preset" later, to make sure that the context is
correctly updated in EngineManagerPrivate::activateEngineItem after
first start of a debugger.

Amends 1c17ed6db8efd767cce99c62f71888dbe55cc1f2

Change-Id: Ib330f06c0f89853b61dc99db05ebe42007ffc48b
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: hjk <hjk@qt.io>