Capitalize the first letter of the types registered to QML
The Qt Quick 2 Extension project wizard allows the users to create
classes with name that starts with lower case letter. Technically it is
correct, but when exporting these classes (with the same name starting
with lower case) they cannot be used from the QML (it will trigger a
"Cannot assign to non-existent property" error). See: http://doc.qt.io/
qt-5/qtqml-typesystem-objecttypes.html
"... the type name must begin with an uppercase letter in order to be
declared as a QML object type in a QML file."

Task-number: QTCREATORBUG-21468
Change-Id: Ic41fce1ae057944886de7458fa275f1b879685dd
Reviewed-by: Miklós Márton <martonmiklosqdev@gmail.com>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>