Fix crash at shutdown if multiple windows are open
The issue is that the window actions were explicitly made children of
the ActionManager, but ActionManager is deleted earlier than the windows
closed at shutdown.

Change WindowList from a bunch of static data to an object with a
lifetime and make it own the actions itself.

Fix-up of 1b0d6e3c26c5dc54c6eae3cb21dc36a22ab1274c which started
deleting the actions in the first place.

Fixes: QTCREATORBUG-21221
Change-Id: I2e335887fa4b85b29bdaa2c908ec643b6abf3231
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>