Squish: Fix tst_CSUP02
Squish (and therefore the test) "is typing" too fast
which breaks some assumption and the (nowadays?) normal
behavior of the CppEditor.
Fix this by introducing a delayedType() function to
be more realistic when doing this test.
Also removes a workaround for QTCREATORBUG-18769 as
this seems no more valid at all.

Task-number: QTCREATORBUG-18679
Change-Id: Ie43554c5b198bf3dc4a26ad12c8368f43bfd21af
Reviewed-by: Christian Stenger <christian.stenger@qt.io>
Reviewed-by: Robert Loehning <robert.loehning@qt.io>