BuildTargetInfo: Make sure changes are actually detected
Make sure that significant changes to the build target information is detected.

Improve operator== to look at all the members. The problem here is that the
runEnvModifier can not meaningfully get compared. So add a uint that is expected
to change whenever the runEnvModifier changes in a significant way.

This makes sure that build target information gets actually updated, which then in
turn enables the updated information to become visible in the UI.

Task-number: QTCREATORBUG-21475
Change-Id: Ifa238f90fc4acba8e29b81959fa40f036b632b3d
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>