Utils: Replace Tooltip hack with another tooltip hack
Instead of fooling moc into believing it operates on a class called
QTipLabel and create a "QTipLabel" class name from that, create a
metaobject with the "QTipLabel" class name directly.

This trades messing around with namespaces for assumptions on
the stability of QMetaObject class layout. Given that this is an
exported class in Qt Core, that's unlikely to break soon.

Change-Id: Iad24dd2a8a7f492e4b238d390888f89c6dd29ea9
Reviewed-by: Eike Ziller <eike.ziller@qt.io>