Tests: Allow removal of objects on the fly
Add another option to allow deletion of unused objects
on the fly.
This option should be used with care. It is recommended
to run without -d (or --delete) beforehand and check the
results.
Anyhow, when deletion is wanted it processes as usual,
renames the original objects.map to objects.map~ and
recreates the objects.map without the objects that
could be deleted.

Change-Id: I3898e8b9998e33461140bf4c75887a32d106f22c
Reviewed-by: Robert Loehning <robert.loehning@qt.io>