Color schemes: Fix wrong colors for unset values
Defaulting to C_TEXT in that case is correct, but the values
for C_TEXT are set via the palette, so other display features like
the current line highlighting and the block visualization,
which are painted _below_ the text, are not overridden.

So, the correct way for a style to set the same foreground and
background colors as C_TEXT, is to leave these colors unset.

Fixes: QTCREATORBUG-21661
Change-Id: I8cdb73e8edc5e3883e1ef107bd1c2b29e6d75b9c
Reviewed-by: Marco Bubke <marco.bubke@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>