MsvcToolChain: Detect Visual Studio installations using vswhere
vswhere is Microsofts recommended way of detecting Visual Studio
installations. Starting with Visual Studio 2017 15.2 it is part of Visual
Studio's installation and according to the documentation can always be
found in %ProgramFiles(x86)%\Microsoft Visual Studio\Installer (see
https://github.com/Microsoft/vswhere).

Given the proper parameters, vswhere is able to find Visual Studio
installions that are older than 2015 (-legacy) and also build tools
(-products *). If the executable is not found, we fall back to the
registry search we did before, but we should use the tool provided
by MS for getting the job of detecting VS installations done.

Change-Id: Ia94e19f00a97138a0261f485c661673e821b06b5
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>