Do not force an artificially high minimum size on the options dialog
We want it to have a reasonable first-use size, but we can just set that
manually when opening it the first time. Afterwards the geometry is
restored from the settings.

All pages already have a vertical scrollbar available, but that also
forced a relatively large minimum size. Reduce this.

Fix-up of dba102ff618496b16c2f3b4772cea7a323946606 and
63be3a4024ff5e4755506295d9e861afd96a9ca6

Fixes: QTCREATORBUG-21678
Change-Id: Iff5810b6ebfe51e3fe4da528bc859598c8df2bb4
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>