SSH: Add ssh-askpass implementation
There is no standard implementation for the ssh-askpass tool on any
platform, so we currently can't find one on our own. In particular, this
means that the RemoteLinux device test will always fail for a newly
added device, which is not acceptable.
So we just add our own askpass implementation and point the SSH settings
to it.

Change-Id: I7ca18725e63f591ef9defdf13768a21cc3de4a54
Reviewed-by: hjk <hjk@qt.io>