Weed out inlined functions when defining trace points
For the allocation functions we only want matched pairs of entry and
return. Otherwise the memory tracking will get confused. Inlined
instances of functions don't have return points, and in general we are
not interested in things that are inlined in libc. We are tracking all
externally visible functions that return allocated memory after all.

(cherry-picked from tqtc-perfprofiler, because of mismatch between
imported and original version)

Change-Id: Ibadb07f8c2e6cabeb473cded7fcc30da8ac2d82d
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>