Fix adding of symbolic trace points
The order of arguments was wrong, and we should really give the
temporary trace point a unique name. The code does not actually work on
a real 32bit libc here as perf cannot properly read the debuginfo, but
it's worth a try in the general case.

(cherry-picked from tqtc-perfprofiler, because of mismatch between
imported and original version)

Change-Id: I769ea396f79cf482d93f1e06fed85e4ed90437d6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>