Additional QNX compatibility checks
cut, df and tail are required for the disk space deploy check.

The QNX code creates PID files in /var/run.

rm is used by the /var/run device test. The QNX Qt Creator code
should be using it to to remove the .pid files that it creates.
Currently, it does not do this.  That'll be a separate fix.

Change-Id: Ie6bbe480b10ad5ac6396e882c993b90c61dcfe70
Reviewed-by: Sami Nurmenniemi <sami.nurmenniemi@qt.io>
Reviewed-by: hjk <hjk@qt.io>