Mode selector: Restrict background painting to selection widget
The FancyTabWidget is spanning the whole window, but only the bar with
the mode buttons and tool buttons should get the background painted.
This fixes that e.g. the titles and labels in the debug mode were dark
instead of having the usual widget background color.

Fix-up of 8770d83fc8bd71691e7cbb5114a1c0b4b8cb4ed8

Change-Id: I9cccb014fd296b9277b675e09a94149b8538a3f1
Reviewed-by: hjk <hjk@qt.io>