Nim: Add code completion with NimSuggest
This patch add basic support for completion by using
the NimSuggest executable.
The user is expected to configure the path to NimSuggest
by using the new tools configuration page in the Nim
settings.
For now we don't support advanced completion of function
parameters. This will be added in further patches.

Change-Id: I5097ee2b3992486071591109c400d9c84da5fc00
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>