QmakeProject: Remove DesktopQmakeRunConfiguration::fromMap
Base functionality is sufficient: The only difference is triggering
updateTargetInformation() and that is triggered (later...) in response
to Project::parsingFinished() too.

The different timing is uncritical as the run configuration is not
accessible in the run settings combo box during that time.

Also, no other runconfigurations use that extra call.

Change-Id: I1c3379ad35baf70d4a45d1a1a2340927eda93785
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>