ClangFormat: Try to add ')' in case of no indentation
clang-format has problems with unfinished () blocks.
In case when the first formatting try failed it makes sense
to try again after appending ')' to the end of the line.

In case it does not solve the initial issue it does not
make it worse anyways.

Examples where this approach helps:

if (a<press CR>

or

if (a <press CR>&& b

Change-Id: Iacc35af79a7b19c1d512d141423d7549788a1a22
Reviewed-by: Marco Bubke <marco.bubke@qt.io>