macOS: Rename profile.icns to qtcreator-project.icns
The profile.icns document type icon is used for all kinds of project
file types which Qt Creator can open.

Let's rename the icon accordingly to something more generic.

The next step is to replace the icon graphics with a version that does
not have the file extension on it.

Change-Id: Ie3b70ae6c8b73e7e414f20ee0a025eb329ee029d
Reviewed-by: Eike Ziller <eike.ziller@qt.io>