Fix that debugger tooltips were overridden by editor tooltips
The debugger tells the editor if the tool tip was handled or not,
but since that information is transferred through a reference
parameter in a signal/slot connection, a second debugger engine
overrode the value of a previous one.
So never set it to 'false' so a previous 'true' is not overridden.

Fixes: QTCREATORBUG-21825
Change-Id: Iec3d15485f224ab0ad14564656886a53fec62e71
Reviewed-by: hjk <hjk@qt.io>