ProjectExplorer: Simplify DeployConfigFactory::availableBuildTargets
It only ever returned either an empty list, or a list with a single
empty string. Using a bool return value to distiguish the cases is
less convoluted.

This is a leftover from the time when a DeployConfigurationFactory
could potentially create more than on DeployConfiguration, but this
is not supported anymore, as factories are cheap nowadays.

Change-Id: I4f2a91d3509d2b978949211ee709863d5ff460c7
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>