CMake: Fix scrolling when editing configuration
Find the index for the new item from the model rather than attempting to
find the item by the row count, which would often return the incorrect
index. It also scrolls to the item so it's in view once it starts editing.

Listen for layout change to scroll to the selected index. This ensures that
when you name an item after adding it, it will scroll to the new location
to edit the value afterward.

Change-Id: Iaa338148a40b921398cfe95da977371a91965a58
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>